+++ flumine/events/events.py
@@ -11,17 +11,21 @@ class EventType(Enum):
     CURRENT_ORDERS = "CurrentOrders"
     CLEARED_MARKETS = "ClearedMarkets"
     CLEARED_ORDERS = "ClearedOrders"
+    CLEARED_ORDERS_META = "ClearedOrders metadata"
+    BALANCE = "Balance"
     # flumine objects
+    STRATEGY = "Strategy"
+    MARKET = "Market"
+    TRADE = "Trade"
+    ORDER = "Order"
     ORDER_PACKAGE = "Order package"
     CLOSE_MARKET = "Closed market"
-    STRATEGY_RESET = "Strategy reset"
     CUSTOM_EVENT = "Custom event"
     NEW_DAY = "New day"
 
 
 class QueueType(Enum):
     HANDLER = "Handler queue"
-    ACCOUNT = "Account queue"
     LOGGING = "Logging queue"
 
 
@@ -39,6 +43,9 @@ class BaseEvent:  # todo __slots__?
 
     def __str__(self):
         return "<{0} [{1}]>".format(self.EVENT_TYPE.name, self.QUEUE_TYPE.name)
+
+
+# HANDLER
 
 
 class MarketCatalogueEvent(BaseEvent):
@@ -76,16 +83,56 @@ class CloseMarketEvent(BaseEvent):
     QUEUE_TYPE = QueueType.HANDLER
 
 
-class StrategyResetEvent(BaseEvent):
-    EVENT_TYPE = EventType.STRATEGY_RESET
-    QUEUE_TYPE = QueueType.HANDLER
-
-
 class CustomEvent(BaseEvent):
     EVENT_TYPE = EventType.CUSTOM_EVENT
     QUEUE_TYPE = QueueType.HANDLER
 
+    def __init__(self, event, callback, *args, **kwargs):
+        super(CustomEvent, self).__init__(event)
+        self.callback = callback
+
 
 class NewDayEvent(BaseEvent):
     EVENT_TYPE = EventType.NEW_DAY
+    QUEUE_TYPE = QueueType.HANDLER
+
+
+# LOGGING
+
+
+class ClearedOrdersMetaEvent(BaseEvent):
+    EVENT_TYPE = EventType.CLEARED_ORDERS_META
+    QUEUE_TYPE = QueueType.LOGGING
+
+
+class BalanceEvent(BaseEvent):
+    EVENT_TYPE = EventType.BALANCE
+    QUEUE_TYPE = QueueType.LOGGING
+
+
+class StrategyEvent(BaseEvent):
+    EVENT_TYPE = EventType.STRATEGY
+    QUEUE_TYPE = QueueType.LOGGING
+
+
+class MarketEvent(BaseEvent):
+    EVENT_TYPE = EventType.MARKET
+    QUEUE_TYPE = QueueType.LOGGING
+
+
+class TradeEvent(BaseEvent):
+    EVENT_TYPE = EventType.TRADE
+    QUEUE_TYPE = QueueType.LOGGING
+
+
+class OrderEvent(BaseEvent):
+    EVENT_TYPE = EventType.ORDER
+    QUEUE_TYPE = QueueType.LOGGING
+
+
+# both
+
+
+class TerminationEvent(BaseEvent):
+    EVENT_TYPE = EventType.TERMINATOR
     QUEUE_TYPE = QueueType.HANDLER