+++ tests/test_middleware.py
@@ -1,7 +1,12 @@
 import unittest
 from unittest import mock
 
-from flumine.markets.middleware import Middleware, SimulatedMiddleware, RunnerAnalytics
+from flumine.markets.middleware import (
+    Middleware,
+    SimulatedMiddleware,
+    RunnerAnalytics,
+    OrderStatus,
+)
 
 
 class MiddlewareTest(unittest.TestCase):
@@ -10,7 +15,7 @@ class MiddlewareTest(unittest.TestCase):
 
     def test_call(self):
         with self.assertRaises(NotImplementedError):
-            self.middleware(None, None)
+            self.middleware(None)
 
 
 class SimulatedMiddlewareTest(unittest.TestCase):
@@ -18,25 +23,49 @@ class SimulatedMiddlewareTest(unittest.TestCase):
         self.middleware = SimulatedMiddleware()
 
     def test_init(self):
-        self.assertEqual(self.middleware.runners, {})
+        self.assertEqual(self.middleware.markets, {})
 
+    @mock.patch(
+        "flumine.markets.middleware.SimulatedMiddleware._process_simulated_orders"
+    )
     @mock.patch("flumine.markets.middleware.SimulatedMiddleware._process_runner")
-    def test_call(self, mock__process_runner):
-        mock_market_catalogue = mock.Mock()
+    def test_call(self, mock__process_runner, mock__process_simulated_orders):
+        mock_market = mock.Mock(context={})
         mock_market_book = mock.Mock()
         mock_runner = mock.Mock()
         mock_runner.status = "ACTIVE"
         mock_market_book.runners = [mock_runner]
-        self.middleware(mock_market_catalogue, mock_market_book)
-        mock__process_runner.assert_called_with(mock_runner)
+        mock_market.market_book = mock_market_book
+        self.middleware(mock_market)
+        mock__process_runner.assert_called_with({}, mock_runner)
+        self.assertEqual(mock_market.context, {"simulated": {}})
+        mock__process_simulated_orders.assert_called_with(mock_market, {})
+
+    def test__process_simulated_orders(self):
+        mock_market_book = mock.Mock()
+        mock_order = mock.Mock()
+        mock_order.status = OrderStatus.EXECUTABLE
+        mock_market = mock.Mock()
+        mock_order_two = mock.Mock()
+        mock_order_two.status = OrderStatus.PENDING
+        mock_order_three = mock.Mock()
+        mock_order_three.status = OrderStatus.EXECUTABLE
+        mock_order_three.simulated = False
+        mock_market.blotter = [mock_order, mock_order_two, mock_order_three]
+        mock_market_analytics = {(mock_order.selection_id, mock_order.handicap): "test"}
+        mock_market.market_book = mock_market_book
+        self.middleware._process_simulated_orders(mock_market, mock_market_analytics)
+        mock_order.simulated.assert_called_with(mock_market_book, "test")
+        mock_order_two.simulated.assert_not_called()
 
     @mock.patch("flumine.markets.middleware.RunnerAnalytics")
     def test__process_runner(self, mock_runner_analytics):
+        market_analytics = {}
         mock_runner = mock.Mock()
-        self.middleware._process_runner(mock_runner)
-        self.assertEqual(len(self.middleware.runners), 1)
-        self.middleware._process_runner(mock_runner)
-        self.assertEqual(len(self.middleware.runners), 1)
+        self.middleware._process_runner(market_analytics, mock_runner)
+        self.assertEqual(len(market_analytics), 1)
+        self.middleware._process_runner(market_analytics, mock_runner)
+        self.assertEqual(len(market_analytics), 1)
         mock_runner_analytics.assert_called_with(mock_runner)
 
 
@@ -59,6 +88,7 @@ class RunnerAnalyticsTest(unittest.TestCase):
             self.runner_analytics._traded_volume, mock_runner.ex.traded_volume
         )
         self.assertEqual(self.runner_analytics.traded, mock__calculate_traded())
+        self.assertEqual(self.runner_analytics._runner, mock_runner)
 
     def test__calculate_traded_dict_empty(self):
         mock_runner = mock.Mock()