+++ tests/test_strategy.py
@@ -122,6 +122,9 @@ class BaseStrategyTest(unittest.TestCase):
     def test_process_orders(self):
         self.strategy.process_orders(None, None)
 
+    def test_process_closed_market(self):
+        self.strategy.process_closed_market(None, None)
+
     def test_finish(self):
         self.strategy.finish()